+++ Modules/Intent.Modules.Common.CSharp/Builder/CSharpClass.cs
@@ -60,6 +60,9 @@ public class CSharpClass : CSharpDeclaration<CSharpClass>, ICSharpClass
         Name = name;
         File = file;
         Parent = parent;
+        
+        _propertiesSeparator = GetConfiguredSeparator(_propertiesSeparator);
+        _fieldsSeparator = GetConfiguredSeparator(_fieldsSeparator);
     }
 
     public CSharpClass(string name) : this(name, Type.Class)
@@ -488,16 +491,16 @@ public class CSharpClass : CSharpDeclaration<CSharpClass>, ICSharpClass
 
     public CSharpClass WithFieldsSeparated(CSharpCodeSeparatorType separator = CSharpCodeSeparatorType.EmptyLines)
     {
-        _fieldsSeparator = separator;
+        _fieldsSeparator = GetConfiguredSeparator(separator);
         return this;
     }
 
     public CSharpClass WithPropertiesSeparated(CSharpCodeSeparatorType separator = CSharpCodeSeparatorType.EmptyLines)
     {
-        _propertiesSeparator = separator;
+        _propertiesSeparator = GetConfiguredSeparator(separator);
         return this;
     }
-
+    
     public CSharpClassMethod? FindMethod(string name)
     {
         return Methods.FirstOrDefault(x => x.Name == name);
@@ -666,7 +669,15 @@ public class CSharpClass : CSharpDeclaration<CSharpClass>, ICSharpClass
     {
         return ToString(indentation);
     }
-
+    
+    private CSharpCodeSeparatorType GetConfiguredSeparator(CSharpCodeSeparatorType defaultSeparator)
+    {
+        var setting = File?.StyleSettings?.BlankLineBetweenMembers;
+        return setting != null && setting.IsAlways()
+            ? CSharpCodeSeparatorType.EmptyLines
+            : defaultSeparator;
+    }
+    
     private CSharpConstructor? GetPrimaryConstructor()
     {
         var primaryConstructors = Constructors.Where(p => p.IsPrimaryConstructor).ToArray();