+++ Modules/Intent.Modules.Common.CSharp/Templates/ICSharpStyleSettings.cs
@@ -20,7 +20,10 @@ public interface ICSharpStyleSettings
     /// </summary>
     ParameterPlacementOptions ParameterPlacement { get; init; }
 
-    BlankLinesBetweenMembersOptions BlankLinesBetweenMembers { get; init; }
+    /// <summary>
+    /// Setting to determine whether to always insert blank lines between members of a class or not
+    /// </summary>
+    BlankLineBetweenMembersOptions BlankLineBetweenMembers { get; init; }
     
     /// <summary>
     /// The order in which elements in the file should be ordered